swarm/swap: own network test file for swap; more tests and bug fixes
