+++ swarm/network_test.go
@@ -21,7 +21,6 @@ import (
 	"flag"
 	"fmt"
 	"io/ioutil"
-	"math/big"
 	"math/rand"
 	"os"
 	"sync"
@@ -375,143 +374,6 @@ func testSwarmNetwork(t *testing.T, o *testSwarmNetworkOptions, steps ...testSwa
 		}
 		log.Debug("done: test sync step", "n", i+1, "nodes", step.nodeCount)
 	}
-}
-
-func TestSwapNetworkSymmetricFileUpload(t *testing.T) {
-	nodeCount := 16
-
-	sim := simulation.New(map[string]simulation.ServiceFunc{
-		"swarm": func(ctx *adapters.ServiceContext, bucket *sync.Map) (s node.Service, cleanup func(), err error) {
-			config := api.NewConfig()
-
-			dir, err := ioutil.TempDir("", "swap-network-test-node")
-			if err != nil {
-				return nil, nil, err
-			}
-			cleanup = func() {
-				err := os.RemoveAll(dir)
-				if err != nil {
-					log.Error("cleaning up swarm temp dir", "err", err)
-				}
-			}
-
-			config.Path = dir
-
-			privkey, err := crypto.GenerateKey()
-			if err != nil {
-				return nil, cleanup, err
-			}
-
-			config.Init(privkey)
-
-			swarm, err := NewSwarm(config, nil)
-			if err != nil {
-				return nil, cleanup, err
-			}
-			bucket.Store(bucketKeySwarm, swarm)
-			log.Info("new swarm", "bzzKey", config.BzzKey, "baseAddr", fmt.Sprintf("%x", swarm.bzz.BaseAddr()))
-			return swarm, cleanup, nil
-		},
-	})
-	defer sim.Close()
-
-	ctx := context.Background()
-	files := make([]file, 0)
-
-	var checkStatusM sync.Map
-	var nodeStatusM sync.Map
-	var totalFoundCount uint64
-
-	_, err := sim.AddNodesAndConnectChain(nodeCount)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	result := sim.Run(ctx, func(ctx context.Context, sim *simulation.Simulation) error {
-		nodeIDs := sim.UpNodeIDs()
-		shuffle(len(nodeIDs), func(i, j int) {
-			nodeIDs[i], nodeIDs[j] = nodeIDs[j], nodeIDs[i]
-		})
-		for _, id := range nodeIDs {
-			key, data, err := uploadFile(sim.Service("swarm", id).(*Swarm))
-			if err != nil {
-				return err
-			}
-			log.Trace("file uploaded", "node", id, "key", key.String())
-			files = append(files, file{
-				addr:   key,
-				data:   data,
-				nodeID: id,
-			})
-		}
-
-		if _, err := sim.WaitTillHealthy(ctx, 2); err != nil {
-			return err
-		}
-
-		// File retrieval check is repeated until all uploaded files are retrieved from all nodes
-		// or until the timeout is reached.
-		for {
-			if retrieve(sim, files, &checkStatusM, &nodeStatusM, &totalFoundCount) == 0 {
-				return nil
-			}
-		}
-	})
-
-	balancesMap := make(map[discover.NodeID]map[discover.NodeID]*big.Int)
-
-	for _, node := range sim.NodeIDs() {
-		item, ok := sim.NodeItem(node, bucketKeySwarm)
-		if !ok {
-			log.Error("No swarm")
-			return
-		}
-		swarm := item.(*Swarm)
-
-		subBalances := make(map[discover.NodeID]*big.Int)
-
-		for _, n := range sim.NodeIDs() {
-			if node == n {
-				continue
-			}
-			balance := swarm.swap.GetPeerBalance(n)
-			if balance != nil {
-				subBalances[n] = balance
-				log.Debug(fmt.Sprintf("Balance of node %s to node %s: %s", node.TerminalString(), n.TerminalString(), swarm.swap.GetPeerBalance(n).String()))
-			} else {
-				log.Debug(fmt.Sprintf("Node %s has no balance with node %s", node.TerminalString(), n.TerminalString()))
-			}
-		}
-		balancesMap[node] = subBalances
-	}
-
-	if *printStats {
-		for k, v := range balancesMap {
-			fmt.Println(fmt.Sprintf("node %s balances:", k.TerminalString()))
-			for kk, vv := range v {
-				fmt.Println(fmt.Sprintf(".........with node %s: balance %s", kk.TerminalString(), vv.String()))
-			}
-		}
-	}
-
-	for k, mapForK := range balancesMap {
-		for n, balanceKwithN := range mapForK {
-			for subK, mapForSubK := range balancesMap {
-				if n == subK {
-					log.Trace(fmt.Sprintf("balance of %s with %s: %s", k.TerminalString(), n.TerminalString(), balanceKwithN))
-					log.Trace(fmt.Sprintf("balance of %s with %s: %s", n.TerminalString(), k.TerminalString(), mapForSubK[k]))
-					if balanceKwithN.CmpAbs(mapForSubK[k]) != 0 && balanceKwithN.Cmp(big.NewInt(0)) != 0 {
-						log.Error("Expected balances to be |abs| = 0 AND balance1 != 0, but they are not")
-					}
-				}
-			}
-		}
-	}
-
-	if result.Error != nil {
-		t.Fatal(result.Error)
-	}
-	log.Debug("test terminated")
 }
 
 // uploadFile, uploads a short file to the swarm instance
